refactor: Split plugin-main
