+++ CMakeLists.txt
@@ -37,6 +37,8 @@ endif()
 target_sources(
   ${CMAKE_PROJECT_NAME}
   PRIVATE src/plugin-main.cpp
+          src/plugin-streaming.cpp
+          src/plugin-stream-recording.cpp
           src/plugin-replay-buffer.cpp
           src/plugin-ui.cpp
           src/utils.cpp
