+++ src/plugin-main.cpp
@@ -24,7 +24,7 @@ with this program. If not, see <https://www.gnu.org/licenses/>
 #include <util/platform.h>
 #include <obs.hpp>
 
-#include <QDateTime>
+#include <QRegularExpression>
 
 #include "audio/audio-capture.hpp"
 #include "video/filter-video-capture.hpp"
@@ -34,9 +34,6 @@ with this program. If not, see <https://www.gnu.org/licenses/>
 
 #define SETTINGS_JSON_NAME "recently.json"
 #define FILTER_ID "osi_branch_output"
-#define OUTPUT_MAX_RETRIES 7
-#define OUTPUT_RETRY_DELAY_SECS 1
-#define RECONNECT_ATTEMPTING_TIMEOUT_NS 2000000000ULL
 #define AVAILAVILITY_CHECK_INTERVAL_NS 1000000000ULL
 #define TASK_INTERVAL_MS 1000
 
@@ -300,35 +297,6 @@ void BranchOutputFilter::destroyCallback()
     obs_log(LOG_INFO, "%s: BranchOutputFilter destroyed", qUtf8Printable(name));
 }
 
-void BranchOutputFilter::stopRecordingOutput(bool pending)
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (recordingOutput) {
-            if (recordingActive) {
-                obs_source_t *parent = obs_filter_get_parent(filterSource);
-                if (parent) {
-                    obs_source_dec_showing(parent);
-                }
-                obs_output_stop(recordingOutput);
-            }
-        }
-        recordingOutput = nullptr;
-
-        if (recordingActive) {
-            recordingActive = false;
-            obs_log(LOG_INFO, "%s: Stopping recording output succeeded", qUtf8Printable(name));
-        }
-
-        recordingPending = pending;
-        recordingSettingsOverridden = false;
-        addChapterToRecordingEnabled = false;
-        splitRecordingEnabled = false;
-    }
-}
-
 void BranchOutputFilter::stopOutput()
 {
     stopRecordingOutput();
@@ -379,133 +347,6 @@ QString BranchOutputFilter::applyFilenameFormatArgs(const QString &format, bool 
     QString filterName = qUtf8Printable(name);
     auto re = noSpace ? QRegularExpression("[\\s/\\\\.:;*?\"<>|&$,]") : QRegularExpression("[/\\\\.:;*?\"<>|&$,]");
     return QString(format).arg(sourceName.replace(re, "-")).arg(filterName.replace(re, "-"));
-}
-
-obs_data_t *BranchOutputFilter::createRecordingSettings(obs_data_t *settings, bool createFolder)
-{
-    auto recordingSettings = obs_data_create();
-    auto config = obs_frontend_get_profile_config();
-
-    // Recording filename format (override takes precedence)
-    QString filenameFormat;
-    if (!recordingFilenameFormatOverride.isEmpty()) {
-        filenameFormat = recordingFilenameFormatOverride;
-    } else {
-        filenameFormat = obs_data_get_string(settings, "filename_formatting");
-        if (filenameFormat.isEmpty()) {
-            filenameFormat = config_get_string(config, "Output", "FilenameFormatting");
-        }
-    }
-
-    // Sanitize filename
-#ifdef __APPLE__
-    filenameFormat.replace(QRegularExpression("[:]"), "");
-#elif defined(_WIN32)
-    filenameFormat.replace(QRegularExpression("[<>:\"\\|\\?\\*]"), "");
-#else
-    // TODO: Add filtering for other platforms
-#endif
-
-    auto useProfileRecordingPath = obs_data_get_bool(settings, "use_profile_recording_path");
-    auto path = useProfileRecordingPath ? getProfileRecordingPath(config) : obs_data_get_string(settings, "path");
-    auto recFormat = obs_data_get_string(settings, "rec_format");
-
-    // Validate recording path
-    if (!path || !path[0]) {
-        obs_log(LOG_ERROR, "%s: Recording path is not set", qUtf8Printable(name));
-        obs_data_release(recordingSettings);
-        return nullptr;
-    }
-
-    if (createFolder) {
-        int ret = os_mkdirs(path);
-        if (ret == MKDIR_ERROR) {
-            obs_log(LOG_ERROR, "%s: Failed to create recording directory: %s", qUtf8Printable(name), path);
-            obs_data_release(recordingSettings);
-            return nullptr;
-        }
-    }
-
-    // Add filter name to filename format
-    bool noSpace = obs_data_get_bool(settings, "no_space_filename");
-    filenameFormat = applyFilenameFormatArgs(filenameFormat, noSpace);
-    auto compositePath = getOutputFilename(path, recFormat, noSpace, false, qUtf8Printable(filenameFormat));
-
-    if (compositePath.isEmpty()) {
-        obs_log(LOG_ERROR, "%s: Recording path is not accessible: %s", qUtf8Printable(name), path);
-        obs_data_release(recordingSettings);
-        return nullptr;
-    }
-
-    obs_data_set_string(recordingSettings, "path", qUtf8Printable(compositePath));
-
-    auto splitFile = obs_data_get_string(settings, "split_file");
-    splitRecordingEnabled = strlen(splitFile) > 0;
-    if (splitRecordingEnabled) {
-        obs_data_set_string(recordingSettings, "directory", path);
-        obs_data_set_string(recordingSettings, "format", qUtf8Printable(filenameFormat));
-        obs_data_set_string(recordingSettings, "extension", qUtf8Printable(getFormatExt(recFormat)));
-        obs_data_set_bool(recordingSettings, "allow_spaces", false);
-        obs_data_set_bool(recordingSettings, "allow_overwrite", false);
-        obs_data_set_bool(recordingSettings, "split_file", true);
-
-        auto maxTimeSec = !strcmp(splitFile, "by_time") ? obs_data_get_int(settings, "split_file_time_mins") * 60 : 0;
-        obs_data_set_int(recordingSettings, "max_time_sec", maxTimeSec);
-
-        auto maxSizeMb = !strcmp(splitFile, "by_size") ? obs_data_get_int(settings, "split_file_size_mb") : 0;
-        obs_data_set_int(recordingSettings, "max_size_mb", maxSizeMb);
-    }
-
-    // Apply fragmented MP4/MOV settings
-    // Immitate https://github.com/obsproject/obs-studio/blob/d3c5d2ce0b15bac7a502f5aef4b3b5ec72ee8e09/frontend/utility/SimpleOutput.cpp#L820
-    QString mux = obs_data_get_string(settings, "rec_muxer_custom");
-    bool isFragmented = strncmp(recFormat, "fragmented", 10) == 0;
-
-    if (isFragmented && (mux.isEmpty() || mux.indexOf("movflags") == -1)) {
-        QString muxFrag = "movflags=frag_keyframe+empty_moov+delay_moov";
-        if (!mux.isEmpty()) {
-            muxFrag += " " + mux;
-        }
-        obs_data_set_string(recordingSettings, "muxer_settings", qUtf8Printable(muxFrag));
-    } else {
-        if (isFragmented) {
-            obs_log(LOG_WARNING, "User enabled fragmented recording, but custom muxer settings contained movflags.");
-        } else {
-            obs_data_set_string(recordingSettings, "muxer_settings", qUtf8Printable(mux));
-        }
-    }
-
-    return recordingSettings;
-}
-
-obs_data_t *BranchOutputFilter::createStreamingSettings(obs_data_t *settings, size_t index)
-{
-    auto streamingSettings = obs_data_create();
-    auto propNameFormat = getIndexedPropNameFormat(index);
-
-    // Copy all settings
-    obs_data_apply(streamingSettings, settings);
-
-    if (index > 0) {
-        // Apply indexed properties (Only on index >= 1 )
-        obs_data_set_string(
-            streamingSettings, "server", obs_data_get_string(settings, qUtf8Printable(propNameFormat.arg("server")))
-        );
-        obs_data_set_string(
-            streamingSettings, "key", obs_data_get_string(settings, qUtf8Printable(propNameFormat.arg("key")))
-        );
-        obs_data_set_bool(
-            streamingSettings, "use_auth", obs_data_get_bool(settings, qUtf8Printable(propNameFormat.arg("use_auth")))
-        );
-        obs_data_set_string(
-            streamingSettings, "username", obs_data_get_string(settings, qUtf8Printable(propNameFormat.arg("username")))
-        );
-        obs_data_set_string(
-            streamingSettings, "password", obs_data_get_string(settings, qUtf8Printable(propNameFormat.arg("password")))
-        );
-    }
-
-    return streamingSettings;
 }
 
 void BranchOutputFilter::getSourceResolution(uint32_t &outWidth, uint32_t &outHeight)
@@ -583,251 +424,6 @@ void BranchOutputFilter::determineOutputResolution(obs_data_t *settings, obs_vid
         ovi->scale_type = OBS_SCALE_BICUBIC;
     } else if (!strcmp(downscaleFilter, "lanczos")) {
         ovi->scale_type = OBS_SCALE_LANCZOS;
-    }
-}
-
-#define FTL_PROTOCOL "ftl"
-#define RTMP_PROTOCOL "rtmp"
-
-BranchOutputFilter::BranchOutputStreamingContext
-BranchOutputFilter::createSreamingOutput(obs_data_t *settings, size_t index)
-{
-    auto count = (size_t)obs_data_get_int(settings, "service_count");
-    if (index >= count || index >= MAX_SERVICES) {
-        return {0};
-    }
-
-    if (!isStreamingEnabled(settings, index)) {
-        return {0};
-    }
-
-    OBSDataAutoRelease streamingSettings = createStreamingSettings(settings, index);
-    BranchOutputStreamingContext context = {0};
-
-    // Create service - We always use "rtmp_custom" as service
-    context.service = obs_service_create("rtmp_custom", qUtf8Printable(name), streamingSettings, nullptr);
-    if (!context.service) {
-        obs_log(LOG_ERROR, "%s: Streaming %zu service creation failed", qUtf8Printable(name), index);
-        return {0};
-    }
-    obs_service_apply_encoder_settings(context.service, streamingSettings, nullptr);
-
-    // Determine output type
-    auto type = obs_service_get_preferred_output_type(context.service);
-    if (!type) {
-        type = "rtmp_output";
-        auto url = obs_service_get_connect_info(context.service, OBS_SERVICE_CONNECT_INFO_SERVER_URL);
-        if (url != nullptr && !strncmp(url, FTL_PROTOCOL, strlen(FTL_PROTOCOL))) {
-            type = "ftl_output";
-        } else if (url != nullptr && strncmp(url, RTMP_PROTOCOL, strlen(RTMP_PROTOCOL))) {
-            type = "ffmpeg_mpegts_muxer";
-        }
-    }
-
-    // Create streaming output
-    context.output =
-        obs_output_create(type, qUtf8Printable(QString("%1 (%2)").arg(name).arg(index)), streamingSettings, nullptr);
-    if (!context.output) {
-        obs_log(LOG_ERROR, "%s (%zu): Streaming output creation failed", qUtf8Printable(name), index);
-        return {0};
-    }
-    obs_output_set_reconnect_settings(context.output, OUTPUT_MAX_RETRIES, OUTPUT_RETRY_DELAY_SECS);
-    obs_output_set_service(context.output, context.service);
-
-    return context;
-}
-
-void BranchOutputFilter::startStreamingOutput(size_t index)
-{
-    if (!streamings[index].output) {
-        return;
-    }
-
-    size_t encIndex = 0;
-    for (size_t i = 0; i < MAX_AUDIO_MIXES; i++) {
-        auto audioContext = &audios[i];
-        if (!audioContext->encoder || !audioContext->streaming) {
-            continue;
-        }
-
-        obs_output_set_audio_encoder(streamings[index].output, audioContext->encoder, encIndex++);
-    }
-
-    if (!encIndex) {
-        // No audio encoder -> fallback first available encoder
-        for (size_t i = 0; i < MAX_AUDIO_MIXES; i++) {
-            auto audioContext = &audios[i];
-            if (audioContext->encoder) {
-                obs_log(
-                    LOG_WARNING, "%s (%zu): No audio encoder selected for streaming output, using track %d",
-                    qUtf8Printable(name), index, i + 1
-                );
-                obs_output_set_audio_encoder(streamings[index].output, audioContext->encoder, encIndex++);
-                break;
-            }
-        }
-        if (!encIndex) {
-            obs_log(LOG_ERROR, "%s (%zu): No audio encoder for streaming output", qUtf8Printable(name), index);
-            return;
-        }
-    }
-
-    obs_output_set_video_encoder(streamings[index].output, videoEncoder);
-
-    streamings[index].outputStarting = true;
-
-    // Track starting signal
-    streamings[index].outputStartingSignal.Connect(
-        obs_output_get_signal_handler(streamings[index].output), "starting",
-        [](void *_data, calldata_t *) {
-            auto context = static_cast<BranchOutputStreamingContext *>(_data);
-            context->outputStarting = true;
-            obs_log(LOG_DEBUG, "%s: Streaming output is starting", obs_output_get_name(context->output));
-        },
-        &streamings[index]
-    );
-
-    // Track activate signal
-    streamings[index].outputActivateSignal.Connect(
-        obs_output_get_signal_handler(streamings[index].output), "activate",
-        [](void *_data, calldata_t *) {
-            auto context = static_cast<BranchOutputStreamingContext *>(_data);
-            context->outputStarting = false;
-            obs_log(LOG_DEBUG, "%s: Streaming output has activated", obs_output_get_name(context->output));
-        },
-        &streamings[index]
-    );
-
-    // Track reconnect signal
-    streamings[index].outputReconnectSignal.Connect(
-        obs_output_get_signal_handler(streamings[index].output), "reconnect",
-        [](void *_data, calldata_t *) {
-            auto context = static_cast<BranchOutputStreamingContext *>(_data);
-            context->reconnectAttemptingAt = os_gettime_ns();
-            obs_log(LOG_DEBUG, "%s: Streaming output is reconnecting", obs_output_get_name(context->output));
-        },
-        &streamings[index]
-    );
-
-    // Track stop signal
-    streamings[index].outputStopSignal.Connect(
-        obs_output_get_signal_handler(streamings[index].output), "stop",
-        [](void *_data, calldata_t *cd) {
-            auto context = static_cast<BranchOutputStreamingContext *>(_data);
-            context->outputStarting = false;
-            auto code = calldata_int(cd, "code");
-            obs_log(
-                LOG_DEBUG, "%s: Streaming output has stopped with code=%lld", obs_output_get_name(context->output), code
-            );
-        },
-        &streamings[index]
-    );
-
-    // Start streaming output
-    if (obs_output_start(streamings[index].output)) {
-        streamings[index].active = true;
-        auto parent = obs_filter_get_parent(filterSource);
-        if (parent) {
-            obs_source_inc_showing(parent);
-        }
-        obs_log(LOG_INFO, "%s (%zu): Starting streaming output succeeded", qUtf8Printable(name), index);
-    } else {
-        obs_log(LOG_ERROR, "%s (%zu): Starting streaming output failed", qUtf8Printable(name), index);
-    }
-}
-
-void BranchOutputFilter::stopStreamingOutput(size_t index)
-{
-    if (streamings[index].output && streamings[index].active) {
-        obs_source_t *parent = obs_filter_get_parent(filterSource);
-        if (parent) {
-            obs_source_dec_showing(parent);
-        }
-        obs_output_stop(streamings[index].output);
-        obs_log(LOG_INFO, "%s (%zu): Stopping streaming output succeeded", qUtf8Printable(name), index);
-    }
-
-    // Ensure signals are disconnected to prevent leaks across restarts
-    streamings[index].outputStartingSignal.Disconnect();
-    streamings[index].outputActivateSignal.Disconnect();
-    streamings[index].outputReconnectSignal.Disconnect();
-    streamings[index].outputStopSignal.Disconnect();
-
-    streamings[index].output = nullptr;
-    streamings[index].service = nullptr;
-    streamings[index].reconnectAttemptingAt = 0;
-    streamings[index].outputStarting = false;
-    streamings[index].active = false;
-    streamings[index].stopping = false;
-}
-
-void BranchOutputFilter::createAndStartRecordingOutput(obs_data_t *settings)
-{
-    if (!videoEncoder) {
-        return;
-    }
-
-    auto recFormat = obs_data_get_string(settings, "rec_format");
-    const char *outputId = !strcmp(recFormat, "hybrid_mp4") ? "mp4_output" : "ffmpeg_muxer";
-
-    // Dtermine chapter marker capability
-    // Chapter maker is only available for hybrid MP4
-    addChapterToRecordingEnabled = !strcmp(recFormat, "hybrid_mp4");
-
-    // Ensure base path exists
-    OBSDataAutoRelease recordingSettings = createRecordingSettings(settings, true);
-    if (!recordingSettings) {
-        obs_log(LOG_ERROR, "%s: Recording settings creation failed (path unavailable?)", qUtf8Printable(name));
-        return;
-    }
-    recordingOutput = obs_output_create(outputId, qUtf8Printable(name), recordingSettings, nullptr);
-    if (!recordingOutput) {
-        obs_log(LOG_ERROR, "%s: Recording output creation failed", qUtf8Printable(name));
-        return;
-    }
-
-    size_t encIndex = 0;
-    for (size_t i = 0; i < MAX_AUDIO_MIXES; i++) {
-        auto audioContext = &audios[i];
-        if (!audioContext->encoder || !audioContext->recording) {
-            continue;
-        }
-
-        obs_output_set_audio_encoder(recordingOutput, audioContext->encoder, encIndex++);
-    }
-
-    if (!encIndex) {
-        // No audio encoder -> fallback first available encoder
-        for (size_t i = 0; i < MAX_AUDIO_MIXES; i++) {
-            auto audioContext = &audios[i];
-            if (audioContext->encoder) {
-                obs_log(
-                    LOG_WARNING, "%s: No audio encoder selected for recording, using track %d", qUtf8Printable(name),
-                    i + 1
-                );
-                obs_output_set_audio_encoder(recordingOutput, audioContext->encoder, encIndex++);
-                break;
-            }
-        }
-        if (!encIndex) {
-            obs_log(LOG_ERROR, "%s: No audio encoder for recording", qUtf8Printable(name));
-            return;
-        }
-    }
-
-    obs_output_set_video_encoder(recordingOutput, videoEncoder);
-
-    // Start recording output
-    if (obs_output_start(recordingOutput)) {
-        recordingActive = true;
-        recordingPending = false;
-        auto parent = obs_filter_get_parent(filterSource);
-        if (parent) {
-            obs_source_inc_showing(parent);
-        }
-        obs_log(LOG_INFO, "%s: Starting recording output succeeded", qUtf8Printable(name));
-    } else {
-        obs_log(LOG_ERROR, "%s: Starting recording output failed", qUtf8Printable(name));
     }
 }
 
@@ -1150,103 +746,6 @@ void BranchOutputFilter::startOutput(obs_data_t *settings)
     }
 }
 
-void BranchOutputFilter::reconnectStreamingOutput(size_t index)
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (streamings[index].active) {
-            obs_output_stop(streamings[index].output);
-
-            if (!obs_output_start(streamings[index].output)) {
-                obs_log(LOG_ERROR, "%s (%zu): Reconnect streaming output failed", qUtf8Printable(name), index);
-            }
-        }
-    }
-}
-
-void BranchOutputFilter::restartRecordingOutput()
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (recordingActive && recordingOutput) {
-            // Re-create recording settings with current override values
-            OBSDataAutoRelease filterSettings = obs_source_get_settings(filterSource);
-            OBSDataAutoRelease newSettings = createRecordingSettings(filterSettings);
-            if (newSettings) {
-                obs_output_update(recordingOutput, newSettings);
-            }
-
-            obs_output_stop(recordingOutput);
-
-            if (!obs_output_start(recordingOutput)) {
-                obs_log(LOG_ERROR, "%s: Restart recording output failed", qUtf8Printable(name));
-            }
-        }
-    }
-}
-
-void BranchOutputFilter::onOverrideRecordingFilenameFormat(void *data, calldata_t *cd)
-{
-    auto filter = static_cast<BranchOutputFilter *>(data);
-
-    const char *format = calldata_string(cd, "format");
-    bool needsSplit = false;
-
-    pthread_mutex_lock(&filter->outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&filter->outputMutex);
-
-        if (!format || !format[0]) {
-            // Empty format -> clear override (revert to filter settings)
-            filter->recordingFilenameFormatOverride.clear();
-            obs_log(LOG_INFO, "%s: Recording filename format override cleared", qUtf8Printable(filter->name));
-        } else {
-            filter->recordingFilenameFormatOverride = QString(format);
-            obs_log(LOG_INFO, "%s: Recording filename format override set to: %s", qUtf8Printable(filter->name), format);
-        }
-
-        // If recording is not active, just store the override for next start
-        if (!filter->recordingActive || !filter->recordingOutput) {
-            return;
-        }
-
-        // Recording is active
-        if (filter->splitRecordingEnabled && !filter->recordingPending) {
-            // Split file enabled: update format setting, then trigger split outside mutex
-            OBSDataAutoRelease filterSettings = obs_source_get_settings(filter->filterSource);
-            bool noSpace = obs_data_get_bool(filterSettings, "no_space_filename");
-            QString appliedFormat = filter->applyFilenameFormatArgs(
-                filter->recordingFilenameFormatOverride.isEmpty()
-                    ? QString(obs_data_get_string(filterSettings, "filename_formatting"))
-                    : filter->recordingFilenameFormatOverride,
-                noSpace
-            );
-
-            OBSDataAutoRelease settings = obs_data_create();
-            obs_data_set_string(settings, "format", qUtf8Printable(appliedFormat));
-            obs_output_update(filter->recordingOutput, settings);
-
-            needsSplit = true;
-            obs_log(
-                LOG_INFO, "%s: Recording output format updated: %s", qUtf8Printable(filter->name),
-                qUtf8Printable(appliedFormat)
-            );
-        } else {
-            // Delegate to intervalTimer: stop (if pending) or restart
-            filter->recordingSettingsOverridden = true;
-        }
-    }
-
-    // Split can be called directly (safe from any thread)
-    if (needsSplit) {
-        filter->splitRecording();
-    }
-}
-
 void BranchOutputFilter::loadProfile(obs_data_t *settings)
 {
     obs_log(LOG_DEBUG, "Profile settings loading");
@@ -1341,104 +840,6 @@ void BranchOutputFilter::restartOutput()
     if (isStreamingGroupEnabled(settings) || isRecordingEnabled(settings) || isReplayBufferEnabled(settings)) {
         startOutput(settings);
     }
-}
-
-bool BranchOutputFilter::reconnectAttemptingTimedOut(size_t index)
-{
-    return streamings[index].reconnectAttemptingAt &&
-           os_gettime_ns() - streamings[index].reconnectAttemptingAt > RECONNECT_ATTEMPTING_TIMEOUT_NS;
-}
-
-bool BranchOutputFilter::someStreamingsStarting()
-{
-    for (size_t i = 0; i < MAX_SERVICES; i++) {
-        if (streamings[i].outputStarting) {
-            return true;
-        }
-    }
-    return false;
-}
-
-int BranchOutputFilter::countEnabledStreamings(obs_data_t *settings)
-{
-    int count = 0;
-    auto serviceCount = (size_t)obs_data_get_int(settings, "service_count");
-    for (size_t i = 0; i < MAX_SERVICES && i < serviceCount; i++) {
-        if (isStreamingEnabled(settings, i)) {
-            count++;
-        }
-    }
-    return count;
-}
-
-bool BranchOutputFilter::isStreamingGroupEnabled(obs_data_t *settings)
-{
-    return obs_data_get_bool(settings, "streaming_enabled") && countEnabledStreamings(settings) > 0;
-}
-
-int BranchOutputFilter::countAliveStreamings()
-{
-    int count = 0;
-    for (size_t i = 0; i < MAX_SERVICES; i++) {
-        if (streamings[i].output && obs_output_active(streamings[i].output)) {
-            count++;
-        }
-    }
-    return count;
-}
-
-int BranchOutputFilter::countActiveStreamings()
-{
-    int count = 0;
-    for (size_t i = 0; i < MAX_SERVICES; i++) {
-        if (streamings[i].active) {
-            count++;
-        }
-    }
-    return count;
-}
-
-bool BranchOutputFilter::hasEnabledStreamings(obs_data_t *settings)
-{
-    for (int i = 0; i < MAX_SERVICES; i++) {
-        if (isStreamingEnabled(settings, i)) {
-            return true;
-        }
-    }
-    return false;
-}
-
-bool BranchOutputFilter::isStreamingEnabled(obs_data_t *settings, size_t index)
-{
-    auto propNameFormat = getIndexedPropNameFormat(index);
-    return !!strlen(obs_data_get_string(settings, qUtf8Printable(propNameFormat.arg("server"))));
-}
-
-bool BranchOutputFilter::isRecordingEnabled(obs_data_t *settings)
-{
-    return obs_data_get_bool(settings, "stream_recording");
-}
-
-bool BranchOutputFilter::isSplitRecordingEnabled(obs_data_t *settings)
-{
-    // Split mode will be disabled when some streamings are enabled.
-    return isRecordingEnabled(settings) && !!strlen(obs_data_get_string(settings, "split_file"));
-}
-
-bool BranchOutputFilter::canPauseRecording()
-{
-    // Pausing recording will be disabled when some streamings are active.
-    return countActiveStreamings() == 0 && !recordingPending;
-}
-
-bool BranchOutputFilter::canAddChapterToRecording()
-{
-    return recordingActive && recordingOutput && addChapterToRecordingEnabled && !obs_output_paused(recordingOutput);
-}
-
-bool BranchOutputFilter::canSplitRecording()
-{
-    return recordingActive && recordingOutput && splitRecordingEnabled && !obs_output_paused(recordingOutput);
 }
 
 // Controlling output status here.
@@ -1745,99 +1146,6 @@ void BranchOutputFilter::onStopOutputGracefully()
 
     // Finalize termination
     stopOutput();
-}
-
-bool BranchOutputFilter::splitRecording()
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (!splitRecordingEnabled || !recordingActive || !recordingOutput) {
-            return false;
-        }
-
-        // Immitate obs_frontend_recording_split_file()
-        proc_handler_t *ph = obs_output_get_proc_handler(recordingOutput);
-        uint8_t stack[128];
-        calldata cd;
-        calldata_init_fixed(&cd, stack, sizeof(stack));
-        proc_handler_call(ph, "split_file", &cd);
-        return calldata_bool(&cd, "split_file_enabled");
-    }
-}
-
-bool BranchOutputFilter::pauseRecording()
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (!recordingActive || !recordingOutput || countActiveStreamings() > 0) {
-            // Can't pause when some streamings are active
-            return false;
-        }
-
-        if (obs_output_paused(recordingOutput)) {
-            // Already paused
-            return false;
-        }
-
-        obs_output_pause(recordingOutput, true);
-        return true;
-    }
-}
-
-bool BranchOutputFilter::unpauseRecording()
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (!recordingActive || !recordingOutput) {
-            return false;
-        }
-
-        if (!obs_output_paused(recordingOutput)) {
-            // Already unpaused
-            return false;
-        }
-
-        obs_output_pause(recordingOutput, false);
-        recordingPending = false; // Force clear pending flag
-        return true;
-    }
-}
-
-bool BranchOutputFilter::addChapterToRecording(QString chapterName)
-{
-    pthread_mutex_lock(&outputMutex);
-    {
-        OBSMutexAutoUnlock locked(&outputMutex);
-
-        if (!addChapterToRecordingEnabled || !recordingActive || !recordingOutput) {
-            return false;
-        }
-
-        if (obs_output_paused(recordingOutput)) {
-            // Can't add chapter when paused
-            return false;
-        }
-
-        // Immitate obs_frontend_recording_add_chapter()
-        proc_handler_t *ph = obs_output_get_proc_handler(recordingOutput);
-        calldata cd;
-        calldata_init(&cd);
-        // Use current date-time when name is empty
-        calldata_set_string(
-            &cd, "chapter_name",
-            chapterName.isEmpty() ? qUtf8Printable(QDateTime::currentDateTime().toString("yyyy-MM-dd hh:mm:ss.zzz"))
-                                  : qUtf8Printable(chapterName)
-        );
-        bool result = proc_handler_call(ph, "add_chapter", &cd);
-        calldata_free(&cd);
-        return result;
-    }
 }
 
 void BranchOutputFilter::setAudioCapturesActive(bool active)
